requirements.txt
bootstrap-flask>=2.2.0
click>=8.1.3
Flask>=2.2.3
gunicorn>=20.1.0
importlib-metadata>=6.1.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.2
numpy>=1.24.2
plotly>=5.13.1
python-dateutil>=2.8.2
scipy>=1.10.1
tenacity>=8.2.2
Werkzeug>=2.2.3
WTForms>=3.0.1
zipp>=3.15.0